requirements.txt
flask
pillow
sklearn
numpy
matplotlib